refactor(controller): 가독성과 관리 용이성을 위해 디렉토리 구조 변경 및 메서드 분리
